+++ src/main/kotlin/controller/Sistem.kt
@@ -1,0 +1,7 @@
+package controller
+
+class Sistem {
+    fun run(){
+        println("Vai dar bom.")
+    }
+}
