+++ src/main/kotlin/dao/GenericDAO.kt
@@ -1,9 +1,10 @@
-package testDriveG.dao
+package dao
 
 interface GenericDAO {
     fun getOne(id : Int) : Any
     fun getAll() : List<Any>
     fun addOne(obj : Any) : Unit
-    fun delete(id : Int) : Any
-    fun update() : Any
+    fun addAll(list : List<Any>) : Unit
+    fun delete(id : Int)
+    fun update(obj : Any)
 }
